# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/base-trips": {
      "post": {
        "operationId": "BaseTripsController_createBaseTrip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBaseTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova Base Trip",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "get": {
        "operationId": "BaseTripsController_findAllBaseTrips",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todas as Base Trips",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-trips/{id}": {
      "get": {
        "operationId": "BaseTripsController_findOneBaseTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhar uma Base Trip",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "patch": {
        "operationId": "BaseTripsController_updateBaseTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBaseTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar uma Base Trip",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "delete": {
        "operationId": "BaseTripsController_removeBaseTrip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar uma Base Trip",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-trips/{id}/days": {
      "post": {
        "operationId": "BaseTripsController_createBaseTripDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBaseTripDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar dia para a Base Trip",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-trip-days/{id}/attractions": {
      "post": {
        "operationId": "BaseTripsController_createBaseAttraction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBaseAttractionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar atração em um dia base",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-attractions/{id}": {
      "patch": {
        "operationId": "BaseTripsController_updateBaseAttraction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBaseAttractionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar atração base",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "delete": {
        "operationId": "BaseTripsController_removeBaseAttraction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover atração base",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-trip-days/{id}/restaurants": {
      "post": {
        "operationId": "BaseTripsController_createBaseRestaurant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBaseRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar restaurante em um dia base",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-restaurants/{id}": {
      "patch": {
        "operationId": "BaseTripsController_updateBaseRestaurant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBaseRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar restaurante base",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "delete": {
        "operationId": "BaseTripsController_removeBaseRestaurant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover restaurante base",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
    "/admin/base-trip-days/{id}": {
      "patch": {
        "operationId": "BaseTripsController_updateBaseTripDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBaseTripDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar dia base",
        "tags": [
          "Admin - Base Trips"
        ]
      },
      "delete": {
        "operationId": "BaseTripsController_removeBaseTripDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover dia base",
        "tags": [
          "Admin - Base Trips"
        ]
      }
    },
...
    "/admin/ai-requests": {
      "get": {
        "operationId": "AdminAiRequestsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tripId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "baseTripId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os logs de requisições de IA",
        "tags": [
          "Admin - AI"
        ]
      }
    },
    "/admin/ai-requests/{id}": {
      "get": {
        "operationId": "AdminAiRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhar um log de requisição de IA",
        "tags": [
          "Admin - AI"
        ]
      }
    },
...
  },
  "info": {
    "title": "2GO / AppRoteiros API",
    "description": "API de Roteiros de Viagem com IA e Autenticação OTP",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBaseTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "numberOfDays": {
            "type": "number"
          },
          "profile": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "fullDescription": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "bestTime": {
            "type": "string"
          },
          "climate": {
            "type": "string"
          },
          "averageBudget": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "INTERNAL"
            ],
            "default": "PRIVATE"
          }
        },
        "required": [
          "title",
          "destination",
          "numberOfDays"
        ]
      },
      "UpdateBaseTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "numberOfDays": {
            "type": "number"
          },
          "profile": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "fullDescription": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "bestTime": {
            "type": "string"
          },
          "climate": {
            "type": "string"
          },
          "averageBudget": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "INTERNAL"
            ],
            "default": "PRIVATE"
          }
        }
      },
      "CreateBaseTripDayDto": {
        "type": "object",
        "properties": {
          "dayNumber": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "suggestedTransport": {
            "type": "string"
          },
          "estimatedCost": {
            "type": "number"
          }
        },
        "required": [
          "dayNumber"
        ]
      },
      "CreateBaseAttractionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "TOURIST_ATTRACTION",
              "MUSEUM",
              "RESTAURANT",
              "CAFE",
              "BAR",
              "BEACH",
              "PARK",
              "SHOPPING",
              "EXPERIENCE",
              "TRANSPORT",
              "EVENT",
              "NIGHTLIFE",
              "FREE_ACTIVITY",
              "PAID_ACTIVITY"
            ]
          },
          "shortDescription": {
            "type": "string"
          },
          "fullDescription": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "googleMapsLink": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "period": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "requiresTicket": {
            "type": "boolean",
            "default": false
          },
          "ticketLink": {
            "type": "string"
          },
          "requiresReservation": {
            "type": "boolean",
            "default": false
          },
          "reservationLink": {
            "type": "string"
          },
          "priority": {
            "type": "number"
          },
          "accessibility": {
            "type": "string"
          },
          "goodForKids": {
            "type": "boolean",
            "default": false
          },
          "goodForElders": {
            "type": "boolean",
            "default": false
          },
          "observations": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "category",
          "order"
        ]
      },
      "UpdateBaseAttractionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "TOURIST_ATTRACTION",
              "MUSEUM",
              "RESTAURANT",
              "CAFE",
              "BAR",
              "BEACH",
              "PARK",
              "SHOPPING",
              "EXPERIENCE",
              "TRANSPORT",
              "EVENT",
              "NIGHTLIFE",
              "FREE_ACTIVITY",
              "PAID_ACTIVITY"
            ]
          },
          "shortDescription": {
            "type": "string"
          },
          "fullDescription": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "googleMapsLink": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "period": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "requiresTicket": {
            "type": "boolean",
            "default": false
          },
          "ticketLink": {
            "type": "string"
          },
          "requiresReservation": {
            "type": "boolean",
            "default": false
          },
          "reservationLink": {
            "type": "string"
          },
          "priority": {
            "type": "number"
          },
          "accessibility": {
            "type": "string"
          },
          "goodForKids": {
            "type": "boolean",
            "default": false
          },
          "goodForElders": {
            "type": "boolean",
            "default": false
          },
          "observations": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
      "CreateBaseRestaurantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cuisineType": {
            "type": "string"
          },
          "priceRange": {
            "type": "string"
          },
          "priceLevel": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "googleMapsLink": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "openingHours": {
            "type": "string"
          },
          "reservationLink": {
            "type": "string"
          },
          "recommendedDish": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "order"
        ]
      },
      "UpdateBaseRestaurantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "cuisineType": {
            "type": "string"
          },
          "priceRange": {
            "type": "string"
          },
          "priceLevel": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "googleMapsLink": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "openingHours": {
            "type": "string"
          },
          "reservationLink": {
            "type": "string"
          },
          "recommendedDish": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
      "UpdateBaseTripDayDto": {
        "type": "object",
        "properties": {
          "dayNumber": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "suggestedTransport": {
            "type": "string"
          },
          "estimatedCost": {
            "type": "number"
          }
        }
      },
...
    }
  }
}